Move vote from `nano/secure` to `nano/lib`
