+++ nano/core_test/message_deserializer.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/blocks.hpp>
+#include <nano/lib/vote.hpp>
 #include <nano/node/transport/message_deserializer.hpp>
-#include <nano/secure/vote.hpp>
 #include <nano/test_common/random.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
