+++ nano/core_test/online_reps.cpp
@@ -1,7 +1,7 @@
+#include <nano/lib/vote.hpp>
 #include <nano/node/election.hpp>
 #include <nano/node/online_reps.hpp>
 #include <nano/node/transport/fake.hpp>
-#include <nano/secure/vote.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
