+++ nano/lib/vote.cpp
@@ -1,7 +1,8 @@
+#include <nano/crypto/blake2/blake2.h>
+#include <nano/lib/object_stream.hpp>
 #include <nano/lib/stream.hpp>
 #include <nano/lib/utility.hpp>
-#include <nano/secure/common.hpp>
-#include <nano/secure/vote.hpp>
+#include <nano/lib/vote.hpp>
 
 #include <boost/property_tree/json_parser.hpp>
