+++ nano/lib/vote.hpp
