+++ nano/node/endpoint.cpp
@@ -1,12 +1,12 @@
 #include <nano/lib/blocks.hpp>
 #include <nano/lib/memory.hpp>
 #include <nano/lib/stream.hpp>
+#include <nano/lib/vote.hpp>
 #include <nano/node/active_elections.hpp>
 #include <nano/node/election.hpp>
 #include <nano/node/endpoint.hpp>
 #include <nano/node/network.hpp>
 #include <nano/node/wallet.hpp>
-#include <nano/secure/vote.hpp>
 
 #include <boost/format.hpp>
