+++ nano/node/local_vote_history.cpp
@@ -1,6 +1,6 @@
+#include <nano/lib/vote.hpp>
 #include <nano/node/local_vote_history.hpp>
 #include <nano/secure/common.hpp>
-#include <nano/secure/vote.hpp>
 
 bool nano::local_vote_history::consistency_check (nano::root const & root_a) const
 {
