+++ nano/node/message_processor.cpp
@@ -1,9 +1,9 @@
 #include <nano/lib/thread_roles.hpp>
+#include <nano/lib/vote.hpp>
 #include <nano/node/bootstrap/bootstrap_service.hpp>
 #include <nano/node/message_processor.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/telemetry.hpp>
-#include <nano/secure/vote.hpp>
 
 nano::message_processor::message_processor (message_processor_config const & config_a, nano::node & node_a) :
 	config{ config_a },
