+++ nano/node/messages.cpp
@@ -7,12 +7,12 @@
 #include <nano/lib/stats_enums.hpp>
 #include <nano/lib/stream.hpp>
 #include <nano/lib/utility.hpp>
+#include <nano/lib/vote.hpp>
 #include <nano/lib/work.hpp>
 #include <nano/node/election.hpp>
 #include <nano/node/endpoint.hpp>
 #include <nano/node/messages.hpp>
 #include <nano/node/network.hpp>
-#include <nano/secure/vote.hpp>
 
 #include <boost/asio/ip/address_v6.hpp>
 #include <boost/endian/conversion.hpp>
