+++ nano/node/vote_rebroadcaster.cpp
@@ -2,6 +2,7 @@
 #include <nano/lib/interval.hpp>
 #include <nano/lib/numbers_templ.hpp>
 #include <nano/lib/thread_roles.hpp>
+#include <nano/lib/vote.hpp>
 #include <nano/node/network.hpp>
 #include <nano/node/rep_tiers.hpp>
 #include <nano/node/vote_processor.hpp>
@@ -9,7 +10,6 @@
 #include <nano/node/vote_router.hpp>
 #include <nano/node/wallet.hpp>
 #include <nano/secure/ledger.hpp>
-#include <nano/secure/vote.hpp>
 
 nano::vote_rebroadcaster::vote_rebroadcaster (nano::vote_rebroadcaster_config const & config_a, nano::ledger & ledger_a, nano::vote_router & vote_router_a, nano::network & network_a, nano::wallets & wallets_a, nano::rep_tiers & rep_tiers_a, nano::stats & stats_a, nano::logger & logger_a) :
 	config{ config_a },
