+++ nano/node/vote_router.cpp
@@ -1,11 +1,11 @@
 #include <nano/lib/enum_util.hpp>
 #include <nano/lib/thread_roles.hpp>
 #include <nano/lib/utility.hpp>
+#include <nano/lib/vote.hpp>
 #include <nano/node/active_elections.hpp>
 #include <nano/node/election.hpp>
 #include <nano/node/vote_cache.hpp>
 #include <nano/node/vote_router.hpp>
-#include <nano/secure/vote.hpp>
 
 #include <chrono>
