+++ nano/secure/fwd.hpp
@@ -13,7 +13,6 @@ class ledger_constants;
 class network_params;
 class pending_info;
 class pending_key;
-class vote;
 
 enum class block_status;
 }
